blog側のサイト構造の変更を修正完了
