--- app/(Blog)/(admin)/admin/layout.tsx
@@ -1,20 +1,0 @@
-import { Metadata } from "next";
-
-export const metadata: Metadata = {
-  title: "ログイン画面",
-  robots: {
-    index: false,
-  },
-};
-
-export default function RootLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return (
-    <>
-      {children}
-    </>
-  );
-}
