+++ app/(Blog)/(blog)/(one-column-page)/sitemaps/page.tsx
@@ -1,10 +1,15 @@
+import { Metadata } from "next";
 import Link from "next/link";
 
-import { getCategories } from "../../lib/service/blogServiceMany";
-import { blogTitle } from "../../config/blogConfig";
+import { getCategories } from "@/app/(blog)/lib/service/blogServiceMany";
+import { blogTitle } from "@/app/(blog)/config/blogConfig";
 
-export const dynamic = "force-static"
+export const dynamic = "force-static";
 export const revalidate = 60 * 60 * 24 * 15;
+
+export const metadata: Metadata = {
+  title: "サイトマップ",
+};
 
 const page = async () => {
   const categories = await getCategories("posts");
@@ -13,13 +18,13 @@ const page = async () => {
     <>
       <h2>サイトマップ</h2>
       <p className="font-semibold text-[#0b0b0c]">
-        <Link href="/memorybook">
+        <Link href="/">
           国内旅行・海外旅行の旅程表作成しおりアプリ「旅のメモリーブック」
         </Link>
       </p>
       <div className="p-4 mb-6 border border-dashed border-blog-borderGray">
         <p className="font-semibold text-[#2a7bdf]">
-          <Link href="/">{blogTitle}</Link>
+          <Link href="/blog">{blogTitle}</Link>
         </p>
         {categories.map((category) => {
           const draftTruePosts = category.posts.filter(
@@ -36,7 +41,10 @@ const page = async () => {
                   <Link href={`/${category.slug}`}>{category.name}</Link>
                 </li>
                 {draftTruePosts.map((post) => (
-                  <li key={post.id} className="my-4 ml-2 md:ml-4 list-disc list-inside">
+                  <li
+                    key={post.id}
+                    className="my-4 ml-2 md:ml-4 list-disc list-inside"
+                  >
                     <Link href={`/${category.slug}/${post.slug}`}>
                       {post.title}
                     </Link>
