+++ app/(Blog)/(dashboard)/dashboard/image/[image_id]/layout.tsx
