+++ app/(Blog)/(dashboard)/dashboard/image/[image_id]/page.tsx
