+++ app/(Blog)/(dashboard)/dashboard/image/new-image/layout.tsx
