+++ app/(Blog)/(dashboard)/dashboard/image/new-image/page.tsx
