+++ app/(Blog)/(dashboard)/dashboard/layout.tsx
